requirements.txt
# ============================
# Core frameworks
# ============================
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.1
pydantic>=2.8.2

# ============================
# Redis client
# ============================
redis>=5.0.7

# ============================
# Qdrant client + LangChain wrapper
# ============================
qdrant-client>=1.9.1
langchain-qdrant>=0.1.2

# ============================
# LangChain core ecosystem
# ============================
langchain>=0.2.10
langchain-core>=0.2.26
langchain-community>=0.2.10
langchain-openai>=0.1.20
openai>=1.37.0

# ============================
# Additional LLM providers
# ============================
# Gemini (Google Generative AI)
langchain-google-genai>=0.1.6
google-generativeai>=0.5.4

# Ollama (local / self-hosted models)
# Solo requiere tener Ollama corriendo, pero añadimos client opcional
ollama>=0.1.9

# ============================
# Extra utilities
# ============================
tqdm>=4.66.5
